+++ NEWS.md
@@ -1,5 +1,16 @@
-# dtq 0.1
+# dtq 0.1.8
 
-- [x] Log all data.table queries in R session memory in separate environment.  
-- [x] `read.only` attribute for data.table.  
-- [x] Measure timing of each query in chain.  
+- [x] rename function dtt to dtl
+- [x] log data.table query in rows
+- [ ] log data.table query out rows (experimental, soon)
+- [x] `read.only` attribute for data.table
+
+
+# dtq 0.1.0
+
+- [x] log data.table query details
+  - [x] call
+  - [x] sequence
+  - [x] environment
+  - [x] timing
+- [x] `read.only` attribute for data.table